Update scrapetube.py
Updated get_ajax_data()
Checking if the response is json or not.  
If the response is not json, provided the response code and attempt to display a message from the server.

In my case, response code 429, with a message parsed from html of:    We're sorry... ... but your computer or network may be sending automated queries. To protect our users, we can't
process your request right now.